+++ examples/oag-file-spec/examples/client-javascript.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/bash
+set -o nounset
+
+cd ../
+make build-javascript test-javascript
+cd examples/
+
+node usage-javascript.js