+++ examples/oag-file-spec/examples/client-python.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/bash
+set -o nounset
+
+cd ../
+make build-python test-python
+cd examples/
+
+node usage_python.py