+++ examples/oag-file-spec/examples/client-ruby.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/bash
+set -o nounset
+
+cd ../
+make build-ruby test-ruby
+cd examples/
+
+node usageruby.rb