+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add run hook, deps extra, and set generator vars
 - Add Markdown lint check
 - Add partial markers
+- Add examples for each component
 
 ### Changed
 - Update Makefile to use MAKEITSO standard